+++ proxy/marshaler_test.go
@@ -19,108 +19,11 @@ package main
 import (
 	"bytes"
 	"encoding/binary"
-	"fmt"
-	"strings"
 	"testing"
-
-	"google.golang.org/protobuf/proto"
 
 	"github.com/google/go-cmp/cmp"
 	gw "github.com/kserve/rest-proxy/gen"
 )
-
-func restRequest(data string, shape string) string {
-	return `{
-	"id": "foo",
-    "parameters": {
-		"top_level": "foo",
-		"bool_param": false
-    },
-	"inputs": [{
-		"name": "predict",
-		"shape": ` + shape + `,
-		"datatype": "FP32",
-		"data":` + data + `,
-        "parameters": {
-            "content_type": "str",
-			"headers": null,
-			"int_param": 42,
-			"bool_param": true
-        }
-		}]
-	}`
-}
-
-var data1D = `
-	[0.0, 0.0, 1.0, 11.0, 14.0, 15.0, 3.0, 0.0, 0.0, 1.0, 13.0, 16.0, 12.0, 16.0, 8.0, 0.0,
-	0.0, 8.0, 16.0, 4.0, 6.0, 16.0, 5.0, 0.0, 0.0, 5.0, 15.0, 11.0, 13.0, 14.0, 0.0, 0.0, 0.0, 0.0,
-	2.0, 12.0, 16.0, 13.0, 0.0, 0.0, 0.0, 0.0, 0.0, 13.0, 16.0, 16.0, 6.0, 0.0, 0.0, 0.0, 0.0, 16.0,
-	16.0, 16.0, 7.0, 0.0, 0.0, 0.0, 0.0, 11.0, 13.0, 12.0, 1.0, 0.0, 0.0, 0.0, 1.0, 11.0, 14.0, 15.0,
-	3.0, 0.0, 0.0, 1.0, 13.0, 16.0, 12.0, 16.0, 8.0, 0.0, 0.0, 8.0, 16.0, 4.0, 6.0, 16.0, 5.0, 0.0,
-	0.0, 5.0, 15.0, 11.0, 13.0, 14.0, 0.0, 0.0, 0.0, 0.0, 2.0, 12.0, 16.0, 13.0, 0.0, 0.0, 0.0, 0.0,
-	0.0, 13.0, 16.0, 16.0, 6.0, 0.0, 0.0, 0.0, 0.0, 16.0, 16.0, 16.0, 7.0, 0.0, 0.0, 0.0, 0.0, 11.0,
-	13.0, 12.0, 1.0, 0.0]
-`
-
-var data2D = `
-[
-	[0.0, 0.0, 1.0, 11.0, 14.0, 15.0, 3.0, 0.0, 0.0, 1.0, 13.0, 16.0, 12.0, 16.0, 8.0,
-	0.0, 0.0, 8.0, 16.0, 4.0, 6.0, 16.0, 5.0, 0.0, 0.0, 5.0, 15.0, 11.0, 13.0, 14.0, 0.0,
-	0.0, 0.0, 0.0, 2.0, 12.0, 16.0, 13.0, 0.0, 0.0, 0.0, 0.0, 0.0, 13.0, 16.0, 16.0, 6.0,
-	0.0, 0.0, 0.0, 0.0, 16.0, 16.0, 16.0, 7.0, 0.0, 0.0, 0.0, 0.0, 11.0, 13.0, 12.0, 1.0, 0.0],
-
-	[0.0, 0.0, 1.0, 11.0, 14.0, 15.0, 3.0, 0.0, 0.0, 1.0, 13.0, 16.0, 12.0, 16.0, 8.0,
-	0.0, 0.0, 8.0, 16.0, 4.0, 6.0, 16.0, 5.0, 0.0, 0.0, 5.0, 15.0, 11.0, 13.0, 14.0, 0.0,
-	0.0, 0.0, 0.0, 2.0, 12.0, 16.0, 13.0, 0.0, 0.0, 0.0, 0.0, 0.0, 13.0, 16.0, 16.0, 6.0,
-	0.0, 0.0, 0.0, 0.0, 16.0, 16.0, 16.0, 7.0, 0.0, 0.0, 0.0, 0.0, 11.0, 13.0, 12.0, 1.0, 0.0]
-]
-`
-
-var data3D = `
-[
-	[
-		[0.0, 0.0, 1.0, 11.0, 14.0, 15.0, 3.0, 0.0, 0.0, 1.0, 13.0, 16.0, 12.0, 16.0, 8.0, 0.0,
-			0.0, 8.0, 16.0, 4.0, 6.0, 16.0, 5.0, 0.0, 0.0, 5.0, 15.0, 11.0, 13.0, 14.0, 0.0, 0.0
-		],
-		[0.0, 0.0, 2.0, 12.0, 16.0, 13.0, 0.0, 0.0, 0.0, 0.0, 0.0, 13.0, 16.0, 16.0, 6.0, 0.0,
-			0.0, 0.0, 0.0, 16.0, 16.0, 16.0, 7.0, 0.0, 0.0, 0.0, 0.0, 11.0, 13.0, 12.0, 1.0, 0.0
-		]
-	],
-	[
-		[0.0, 0.0, 1.0, 11.0, 14.0, 15.0, 3.0, 0.0, 0.0, 1.0, 13.0, 16.0, 12.0, 16.0, 8.0, 0.0,
-			0.0, 8.0, 16.0, 4.0, 6.0, 16.0, 5.0, 0.0, 0.0, 5.0, 15.0, 11.0, 13.0, 14.0, 0.0, 0.0
-		],
-		[0.0, 0.0, 2.0, 12.0, 16.0, 13.0, 0.0, 0.0, 0.0, 0.0, 0.0, 13.0, 16.0, 16.0, 6.0, 0.0,
-			0.0, 0.0, 0.0, 16.0, 16.0, 16.0, 7.0, 0.0, 0.0, 0.0, 0.0, 11.0, 13.0, 12.0, 1.0, 0.0
-		]
-	]
-]
-`
-
-var data4D = `
-[
-	[
-		[
-			[0.0, 0.0, 1.0, 11.0, 14.0, 15.0, 3.0, 0.0, 0.0, 1.0, 13.0, 16.0, 12.0, 16.0, 8.0, 0.0],
-			[0.0, 8.0, 16.0, 4.0, 6.0, 16.0, 5.0, 0.0, 0.0, 5.0, 15.0, 11.0, 13.0, 14.0, 0.0, 0.0]
-		],
-		[
-			[0.0, 0.0, 2.0, 12.0, 16.0, 13.0, 0.0, 0.0, 0.0, 0.0, 0.0, 13.0, 16.0, 16.0, 6.0, 0.0],
-			[0.0, 0.0, 0.0, 16.0, 16.0, 16.0, 7.0, 0.0, 0.0, 0.0, 0.0, 11.0, 13.0, 12.0, 1.0, 0.0]
-		]
-	],
-	[
-		[
-			[0.0, 0.0, 1.0, 11.0, 14.0, 15.0, 3.0, 0.0, 0.0, 1.0, 13.0, 16.0, 12.0, 16.0, 8.0, 0.0],
-			[0.0, 8.0, 16.0, 4.0, 6.0, 16.0, 5.0, 0.0, 0.0, 5.0, 15.0, 11.0, 13.0, 14.0, 0.0, 0.0]
-		],
-		[
-			[0.0, 0.0, 2.0, 12.0, 16.0, 13.0, 0.0, 0.0, 0.0, 0.0, 0.0, 13.0, 16.0, 16.0, 6.0, 0.0],
-			[0.0, 0.0, 0.0, 16.0, 16.0, 16.0, 7.0, 0.0, 0.0, 0.0, 0.0, 11.0, 13.0, 12.0, 1.0, 0.0]
-		]
-	]
-]
-`
 
 func generateProtoBufResponse() *gw.ModelInferResponse {
 	expectedOutput := []*gw.ModelInferResponse_InferOutputTensor{{
@@ -147,60 +50,6 @@ func generateProtoBufResponse() *gw.ModelInferResponse {
 
 var jsonResponse = `{"model_name":"example","id":"foo","parameters":{"bool_param":false,"content_type":"bar","headers":null,"int_param":12345},` +
 	`"outputs":[{"name":"predict","datatype":"INT64","shape":[2],"data":[8,8]}]}`
-
-func generateProtoBufRequest(shape []int64) *gw.ModelInferRequest {
-	var expectedInput = gw.ModelInferRequest_InferInputTensor{
-		Name:     "predict",
-		Datatype: "FP32",
-		Shape:    shape,
-		Parameters: map[string]*gw.InferParameter{
-			"content_type": {ParameterChoice: &gw.InferParameter_StringParam{StringParam: "str"}},
-			"headers":      {ParameterChoice: nil},
-			"int_param":    {ParameterChoice: &gw.InferParameter_Int64Param{Int64Param: 42}},
-			"bool_param":   {ParameterChoice: &gw.InferParameter_BoolParam{BoolParam: true}},
-		},
-		Contents: &gw.InferTensorContents{
-			Fp32Contents: []float32{0.0, 0.0, 1.0, 11.0, 14.0, 15.0, 3.0, 0.0, 0.0, 1.0, 13.0, 16.0, 12.0,
-				16.0, 8.0, 0.0, 0.0, 8.0, 16.0, 4.0, 6.0, 16.0, 5.0, 0.0, 0.0, 5.0, 15.0, 11.0, 13.0,
-				14.0, 0.0, 0.0, 0.0, 0.0, 2.0, 12.0, 16.0, 13.0, 0.0, 0.0, 0.0, 0.0, 0.0, 13.0, 16.0,
-				16.0, 6.0, 0.0, 0.0, 0.0, 0.0, 16.0, 16.0, 16.0, 7.0, 0.0, 0.0, 0.0, 0.0, 11.0, 13.0,
-				12.0, 1.0, 0.0, 0.0, 0.0, 1.0, 11.0, 14.0, 15.0, 3.0, 0.0, 0.0, 1.0, 13.0, 16.0, 12.0,
-				16.0, 8.0, 0.0, 0.0, 8.0, 16.0, 4.0, 6.0, 16.0, 5.0, 0.0, 0.0, 5.0, 15.0, 11.0, 13.0,
-				14.0, 0.0, 0.0, 0.0, 0.0, 2.0, 12.0, 16.0, 13.0, 0.0, 0.0, 0.0, 0.0, 0.0, 13.0, 16.0,
-				16.0, 6.0, 0.0, 0.0, 0.0, 0.0, 16.0, 16.0, 16.0, 7.0, 0.0, 0.0, 0.0, 0.0, 11.0, 13.0,
-				12.0, 1.0, 0.0},
-		},
-	}
-
-	var modelInferRequest = &gw.ModelInferRequest{
-		Id: "foo",
-		Parameters: map[string]*gw.InferParameter{
-			"top_level":  {ParameterChoice: &gw.InferParameter_StringParam{StringParam: "foo"}},
-			"bool_param": {ParameterChoice: &gw.InferParameter_BoolParam{BoolParam: false}},
-		},
-		Inputs: []*gw.ModelInferRequest_InferInputTensor{&expectedInput},
-	}
-	return modelInferRequest
-}
-
-func TestRESTRequest(t *testing.T) {
-	c := CustomJSONPb{}
-	inputDataArray := []string{data1D, data2D, data3D, data4D}
-	inputDataShapes := [][]int64{{2, 64}, {2, 64}, {2, 2, 32}, {2, 2, 2, 16}}
-	for k, data := range inputDataArray {
-		out := &gw.ModelInferRequest{}
-		buffer := &bytes.Buffer{}
-		buffer.Write([]byte(restRequest(data, strings.Join(strings.Split(fmt.Sprintln(inputDataShapes[k]), " "), ","))))
-		err := c.NewDecoder(buffer).Decode(out)
-		if err != nil {
-			t.Error(err)
-		}
-		expected := generateProtoBufRequest(inputDataShapes[k])
-		if !proto.Equal(out, expected) {
-			t.Errorf("REST request failed to decode for shape: %v", inputDataShapes[k])
-		}
-	}
-}
 
 func TestRESTResponse(t *testing.T) {
 	c := CustomJSONPb{}
